requirements.txt
Django==4.1.5
django_phonenumber_field==7.0.2
mysqlclient==2.1.1